requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
croniter==1.3.8
Deprecated==1.2.13
Flask==2.2.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack-python==0.5.6
psutil==5.9.4
pycparser==2.21
PyGithub==1.58.0
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
